# Remove HashProvider and deprecate it