--- a/src/main/php/util/collections/DJBX33AHashImplementation.class.php
+++ b/src/main/php/util/collections/DJBX33AHashImplementation.class.php
@@ -31,6 +31,7 @@
  *
  * -- Ralf S. Engelschall
  *
+ * @deprecated
  * @see  xp://util.collections.HashProvider
  * @test xp://net.xp_framework.unittest.util.collections.DJBX33AHashImplementationTest:
  */