--- a/src/main/php/util/collections/HashImplementation.class.php
+++ b/src/main/php/util/collections/HashImplementation.class.php
@@ -3,6 +3,7 @@
 /**
  * Hash implementation
  *
+ * @deprecated
  * @see   xp://util.collections.HashProvider
  */
 interface HashImplementation {