--- a/src/main/php/util/collections/MD5HashImplementation.class.php
+++ b/src/main/php/util/collections/MD5HashImplementation.class.php
@@ -3,6 +3,7 @@
 /**
  * MD5
  *
+ * @deprecated
  * @see   php://md5
  * @see   xp://util.collections.HashProvider
  * @test  xp://net.xp_framework.unittest.util.collections.MD5HashImplementationTest: