This pull request changes the way hashcodes are computed for `HashTable`, `HashSet`, `Queue` and `Stack`. Instead of keeping track of the hashcode when writing, it is computed when `hashCode()` is invoked, which occurs much less often!

This yields a speedup factor between 2 and 4 - see [this measuring code](https://gist.github.com/thekid/4a3e9f5b51aa140218468ca8b922d223):

```bash
# Before
$ xp -m ..\measure\ util.profiling.Measure Collections -n 100000
hashTable: 100000 iteration(s), 0.674 seconds, result= "value"
hashSet: 100000 iteration(s), 0.458 seconds, result= 1
queue: 100000 iteration(s), 0.576 seconds, result= "green"
stack: 100000 iteration(s), 0.404 seconds, result= "green"

# After
$ xp -m ..\measure\ util.profiling.Measure Collections -n 100000
hashTable: 100000 iteration(s), 0.332 seconds, result= "value"
hashSet: 100000 iteration(s), 0.249 seconds, result= 1
queue: 100000 iteration(s), 0.135 seconds, result= "green"
stack: 100000 iteration(s), 0.097 seconds, result= "green"
```

This pull request does not introduce a BC break. However, future major versions will remove the now-deprecated class `HashProvider`. Although it's a purely *internal* class is was not clearly marked as such and might have gotten used.